% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{search_organization_multiple}
\alias{search_organization_multiple}
\title{Search Organization (Multiple)}
\usage{
search_organization_multiple(
  term,
  address,
  yelp_api_key,
  num_responses,
  lat = NULL,
  lon = NULL,
  radius = NULL
)
}
\arguments{
\item{term}{A string specifying the search term (e.g., business name, category).}

\item{address}{A string representing the location address.}

\item{yelp_api_key}{the Yelp API key for authentication.}

\item{lat}{Optional value representing the latitude of the location. If this parameter is provided, address should not be provided.}

\item{lon}{Optional value representing the longitude of the location. If this parameter is provided, address should not be provided.}

\item{radius}{Optional value representing a maximum search radius in miles}
}
\value{
A data table of information about each matching search result
}
\description{
This function searches for a organizations matching a search term near a location
}
