% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{search_organization_single}
\alias{search_organization_single}
\title{Search Organization (Single)}
\usage{
search_organization_single(term, address, yelp_api_key, lat = NULL, lon = NULL)
}
\arguments{
\item{term}{A string specifying the search term (e.g., business name, category).}

\item{address}{A string representing the location address.}

\item{yelp_api_key}{the Yelp API key for authentication.}

\item{lat}{Optional value representing the latitude of the location. If this parameter is provided, address should not be provided.}

\item{lon}{Optional value representing the longitude of the location. If this parameter is provided, address should not be provided.}
}
\value{
A vector containing the review count, rating, and phone number of the found organization.
If the organization is not found or an error occurs, NULL is returned.
}
\description{
This function searches for a single organization based on a search term and address using the Yelp Fusion API.
}
